project

+++ pallets/department-funding/src/lib.rs
@@ -41,7 +41,7 @@ type AccountIdOf<T> = <T as frame_system::Config>::AccountId;
 type BalanceOf<T> = <<T as Config>::Currency as Currency<AccountIdOf<T>>>::Balance;
 pub type BlockNumberOf<T> = <T as frame_system::Config>::BlockNumber;
 pub type SumTreeNameType<T> = SumTreeName<AccountIdOf<T>, BlockNumberOf<T>>;
-type DeparmentId = u128;
+type DeparmentId = u64;
 
 #[frame_support::pallet]
 pub mod pallet {
